finish all the php files and js fetch request

+++ api.php
@@ -1,18 +1,28 @@
 <?php
 
     require_once('API/DB.php');
+    require_once('API/HTTP.php');
 
-    $arr = $_POST;
+    $action = $_GET['action'];
+
+    // "GETTING" and "POSTTING" the data
+    if($action == 'get')
+    {
+        $arr = array();
+        parse_str($_SERVER['QUERY_STRING'], $arr);
+    }
+    else
+    {
+        $arr = $_POST;
+
+    }
     $array = array();
 
+    // order the data in values
     foreach ($arr as $key => $value) 
     {
         switch ($key) 
-        {
-            case 'action':
-                $action = $value;
-            break;
-            
+        {            
             case 'table':
                 $table = $value;
             break;
@@ -30,31 +40,67 @@
         }
     }
 
+    // calling the DB functions according the action + validate empty data
     switch ($action) {
         case 'insert':
+
+            if(empty($table) || empty($array))
+            {
+                $msg = "The server could not understand the request due to invalid syntax.";
+                HTTP::create('400' , $msg);
+                break;
+            }
+
             $db = new DB();
             $msg = $db->insert($table, $array);
-            echo $msg;
+            HTTP::create('201' , $msg);
             break;
         
         case 'delete':
+
+            if(empty($table) || empty($field) || empty($data))
+            {
+                $msg = "The server could not understand the request due to invalid syntax.";
+                HTTP::create('400' , $msg);
+                break;
+            }
+
             $db = new DB();
             $msg = $db->delete($table, $field, $data);
-            echo $msg;
+            HTTP::create('200' , $msg);
             break;
         
         case 'get':
+
+            if(empty($table) || empty($field) || empty($data))
+            {
+                $msg = "The server could not understand the request due to invalid syntax.";
+                HTTP::create('400' , $msg);
+                break;
+            }
+
             $db = new DB();
             $msg = $db->get($table, $field, $data);
-            print_r($msg);
+            HTTP::create('200' , $msg);
             break;
         
         case 'update':
+
+            if(empty($table) || empty($field) || empty($data) || empty($array))
+            {
+                $msg = "The server could not understand the request due to invalid syntax.";
+                HTTP::create('400' , $msg);
+                break;
+            }
+
             $db = new DB();
             $msg = $db->update($table, $field, $data, $array);
-            echo $msg;
+            HTTP::create('200' , $msg);
             break;
+            
         default:
-            echo 'diff';
+            $msg = "The server could not understand the request due to invalid syntax.";
+            HTTP::create('400' , $msg);
+            break;
       }
 ?>
